# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: НейроКлип API — AI video production platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "200":
          description: Registered + session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/oauth/{provider}:
    post:
      operationId: oauthLogin
      tags: [auth]
      summary: Mock OAuth login (VK / Yandex)
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [vk, yandex]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
...
  /billing/subscribe:
    post:
      operationId: subscribeToPlan
      tags: [billing]
      summary: Subscribe to plan (mock — returns confirmation_url)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscribeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentConfirmation" }
  /billing/tokens/purchase:
    post:
      operationId: purchaseTokens
      tags: [billing]
      summary: Purchase token pack (mock — instantly credits balance)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseTokensRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentConfirmation" }
...
  /billing/promo:
    post:
      operationId: redeemPromoCode
      tags: [billing]
      summary: Redeem promo code (adds tokens or discount note)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromoCodeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PromoCodeResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    RegisterRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
        consent: { type: boolean, description: "ФЗ-152 consent" }
      required: [email, password, name, consent]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/CurrentUser" }
      required: [user]
...
    CurrentUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        role: { type: string, enum: [user, admin] }
        planId: { type: string }
        planName: { type: string }
        tokenBalance: { type: integer }
        videosUsedThisPeriod: { type: integer }
        videosQuota: { type: integer }
        currentPeriodEnd: { type: ["string", "null"], format: date-time }
        cancelAtPeriodEnd: { type: boolean }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - email
        - name
        - role
        - planId
        - planName
        - tokenBalance
        - videosUsedThisPeriod
        - videosQuota
        - cancelAtPeriodEnd
        - createdAt
...
    SubscribeRequest:
      type: object
      properties:
        planId: { type: string }
        period: { type: string, enum: [month, year] }
      required: [planId, period]
    PurchaseTokensRequest:
      type: object
      properties:
        pack: { type: string, enum: [tokens_100, tokens_500, tokens_2000] }
      required: [pack]
    PaymentConfirmation:
      type: object
      properties:
        paymentId: { type: string }
        confirmationUrl: { type: string }
        amountRub: { type: number }
        status: { type: string }
      required: [paymentId, confirmationUrl, amountRub, status]
...
    PromoCodeRequest:
      type: object
      properties:
        code: { type: string }
      required: [code]
    PromoCodeResponse:
      type: object
      properties:
        ok: { type: boolean }
        message: { type: string }
        tokensAdded: { type: integer }
      required: [ok, message, tokensAdded]
...